방문 인증 api 배포

+++ src/main/java/com/favoriteplace/app/dto/travel/PilgrimageDto.java
@@ -98,7 +98,7 @@ public class PilgrimageDto {
     @NoArgsConstructor(access = PROTECTED)
     @AllArgsConstructor(access = PRIVATE)
     public static class PilgrimageCertifyRequestDto{
-        Long longitude;
-        Long latitude;
+        Double longitude;
+        Double latitude;
     }
 }
